Refactored Length class.

+++ core/src/main/java/com/spinyowl/spinygui/core/style/types/length/LengthConverter.java
@@ -15,7 +15,7 @@ public interface LengthConverter<T extends Number> {
    * @param baseLength base length (that could be used to calculate relative length).
    * @return calculated length.
    */
-  int convert(Length<T> original, float baseLength);
+  float convert(Length<T> original, float baseLength);
 
   /**
    * Used to convert length of specified type to pixels. Base length is 1.
@@ -23,7 +23,7 @@ public interface LengthConverter<T extends Number> {
    * @param original original length.
    * @return calculated length.
    */
-  default int convert(Length<T> original) {
+  default float convert(Length<T> original) {
     return convert(original, 1);
   }
 }
